+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Version 0.1.1
 
 Add least squares linear regression method for single or multivariate input; specifically the ``QaQc.lin_regress()`` method. It can be used to correct energy balance components or for any arbitrary time series data loaded in a ``QaQc`` instance. It produces and returns a readable table with regression results (fitted coefficients, root-mean-square-error, etc.) which can be accessed from ``QaQc.lin_regress_results`` after calling the method. The default regression if used to correct energy balance components assumes net radiation is accurate (as the dependent variable):
 
-:math:`Rn = c_0 - c_1 G - c_2 LE - c_3 H`
+:math:`Rn = c_0 + c_1 G + c_2 LE + c_3 H`
 
 where :math:`c_0 = 0`.